Claude Apps/python-ai-roadmap/phase3/module-m/pyproject.toml
[project]
name = "module-m"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.30.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "streamlit>=1.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
